+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Added
 
+- BaseAvatar: Add micro size.
+
 - BaseDivider: Add margin prop.
 
 - FieldCheckbox: Add autofocus, error, info, success, validation and warning prop.